gallery update
added carousel_slider to package and corrected the bug of gallery page scroll
