+++ lib/packages/carousel_slider/carousel_options.dart
