+++ lib/packages/carousel_slider/carousel_slider.dart
