+++ lib/packages/carousel_slider/utils.dart
