+++ lib/screens/home_screen/pages/HomePage/widgets/carousel.dart
@@ -1,4 +1,4 @@
-import 'package:cssapp/widgets/carousel_slider/carousel_slider.dart';
+import 'package:cssapp/packages/carousel_slider/carousel_slider.dart';
 
 import 'package:flutter/material.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
